fix: scheduled verification is not run in transactional context
In Spring, self-invocations do not start new transactional contexts. The `@Scheduled` method `tryVerifyAllPending` invokes `verify(LoggingContext, GAVCPE)` on the same instance - a self-invocation. However, the scheduled method is not run in a transactional context. Thus, despite the `@Transactional` annotation, this might is called without a transactional context in this specific case.

This change fixes this by also requiring a `@Transactional` context for `tryVerifyAllPending`. While the self-invocation still does not start a new context, it will continue to run in the now existing context created for the scheduled method.

Co-authored-by: Felix <23635466+its-felix@users.noreply.github.com>